add CollisionBox and some basic UI

+++ src/math/Position.java
@@ -5,7 +5,7 @@ import mechanic.Movement;
 public class Position {
     private double x;
     private double y;
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
